refactor: move photos screen logic to features module
- Move PhotosScreen component from app/(tabs) to features/photos/components
- Refactor photos.tsx to import and render PhotosScreen from features module
- Fix circular dependency by using direct imports instead of barrel exports
- Export PhotosScreen from features/photos components index

This improves code organization by keeping screen logic in the features module
while maintaining clean separation between routing and business logic.
