+++ src/features/photos/components/index.ts
@@ -1,3 +1,4 @@
 export { PhotoItem } from './photo-item';
 export { PhotoGrid } from './photo-grid';
 export { PhotoDetailItem } from './photo-detail-item';
+export { PhotosScreen } from './photos-screen';
